Fix election presenter namespace
