+++ decidim-elections/app/models/decidim/elections/election.rb
@@ -16,7 +16,7 @@ module Decidim
       has_many :questions, foreign_key: "decidim_elections_election_id", class_name: "Decidim::Elections::Question", inverse_of: :election, dependent: :destroy
 
       def self.log_presenter_class_for(_log)
-        Decidim::AdminLog::ElectionPresenter
+        Decidim::Elections::AdminLog::ElectionPresenter
       end
 
       def started?
